venta por peso: unidad g/kg en productos, POS y compras
- Productos pueden configurarse como Unidades / Gramos / Kilogramos
- POS: tarjeta de producto con selector de peso + preview de precio
- POS sidebar: input decimal con etiqueta de unidad para ítems de peso
- OrderService/StockService soportan qty float (decimal 10,3)
- Pestaña "Compra de producto" en vista de compras con selector de unidad
- UnitConverter: soporte para base kg (factorToBase)
- Tests: UnitConverterTest (7) + WeightProductTest (4), todos verdes
- Migraciones base para entorno de test (core_pos_tables, supply_purchases)

+++ app/Models/Order.php
@@ -272,7 +272,7 @@ class Order extends Model
                     'quantity' => $item->quantity
                 ]);
 
-                $stock->adjust($product, - (int) $item->quantity, 'venta', $this);
+                $stock->adjust($product, - (float) $item->quantity, 'venta', $this);
 
                 // Descontar insumos del producto
                 $this->reduceSuppliesForProduct($product, $item->quantity);
